Add field "pid" if no TCA columns configuration has been set for selected table
While field "pid" has been configured in TCA columns for table "tt_address", this is not the case for other tables like "pages" or "tt_content". In these cases the field array for the fields available for selection is extended by the field "pid" to make it available for selection when uploading data.