[TASK] Adopt latest `php-cs-fixer` configuration from TYPO3
The TYPO3 development is always a good inspiration to follow
also for project and/or extension development and considerable
a good practice to follow.

This change adopts the latest TYPO3 configuration for the
`php-cs-fixer` related to PHP coding style expect for one
case. TYPO3 uses `self::` in phpunit tests to call methods
definest as static method, albeit phpunit and the related
maintainer (Sebastian Bergman) recommends to acces these
methods using `$this->`. Galdly, there is a php-cs-fixer
rule for that which is configured other then for TYPO3.

The TYPO3 Core stays away from changeing that without any
good reasoning due to the huge amount of unit and functional
tests that would change and special for backports not making
it worth to invest reviewer time on that. For extensions and
special for this one this is more overseable so let's change
that.

> [!NOTE]
> This change is only the rule configuration leading to
> reported changes, which will has been applied and added
> as a pre-commit.