[BUGFIX] Hide upload form when no allowed tables are configured
When either no `allowedTables` are configured using the documented
`PageTSConfig`, either globally or for the page currently selected,
it does not make sense to provide the import file upload form as it
is not possible to guess any table it should be imported to and lead
to follow up issues.

This change modifies the fluid template checking the count of the
available tables, displaying a error message instead of the upload
form if no tables are allowed. Otherwise the upload form with the
select field for the table is rendered, moved to dedicated section
now.