[TASK] Convert ValueObjects during import to plain string values
In some cases a spreadsheet can contain complex field values like RichText or Formulas, which are represented by the library as ValueObjects.

We now check for objects in the parsed field values, and check if they can be converted to strings.

If they can not be converted a flashmessge will be displayed and the value will be represented as 'N/A'