[TASK] Prepare `master` branch for next major `7.x`
After a short internal discussion the decision has been
made to have `6.0.0` as only `6.x` version and phase it
out and work towards new major `7.x`.

Main reason for this decision is the fact, that it has
been missed to remove deprecated functionality for the
`6.0.0` release, which cannot be done not as `6.x` due
to beeing breaking and violating `semver`.

Used command(s):

```shell
Build/Scripts/runTests.sh -p 8.2 -t 13 \
  -s composer -- config \
    "extra"."branch-alias"."dev-master" "7.x-dev"
```